+++ Utilities/MathUtilities.cs
@@ -6,29 +6,29 @@ namespace WallpaperGenerator.Utilities
 {
     public static class MathUtilities
     {
-        public static double MathExpectation(double[] values)
+        public static double MathExpectation(float[] values) 
         {
             double sum = values.Sum();
             return sum / values.Length;
         }
 
-        public static double Variance(double[] values)
+        public static double Variance(float[] values)
         {
             if (values.Length == 1)
                 return 0;
 
             double sum = values.Sum();
-            double sumOfSquares = values.Sum(t => t*t);
+            double sumOfSquares = values.Sum(t => t * t);
 
             return (sumOfSquares - sum * sum / values.Length) / (values.Length - 1);
         }
 
-        public static double StandardDeviation(double[] values)
+        public static double StandardDeviation(float[] values)
         {
             double varianse = Variance(values);
             return Math.Sqrt(varianse);
         }
-        
+
         public static double Map(double value, double rangeStart, double rangeEnd, double range,
             double mappedRangeStart, double mappedRangeEnd, double mappedRange, double scale)
         {
